rl/algos/ppo_checkpointing.py
from __future__ import annotations

import json
from pathlib import Path

import torch


def _append_jsonl(path: Path, record: dict) -> None:
    path.parent.mkdir(parents=True, exist_ok=True)
    with path.open("a", encoding="utf-8") as f:
        f.write(json.dumps(record, sort_keys=True) + "\n")


def _checkpoint_paths(exp_dir: Path, iteration: int) -> tuple[Path, Path]:
    ckpt_dir = exp_dir / "checkpoints"
    ckpt_dir.mkdir(parents=True, exist_ok=True)
    ckpt_path = ckpt_dir / f"iter_{int(iteration):06d}.pt"
    latest_path = exp_dir / "checkpoint_last.pt"
    return ckpt_path, latest_path


def _save_checkpoint(path: Path, state: dict) -> None:
    path.parent.mkdir(parents=True, exist_ok=True)
    torch.save(state, path)


def _load_checkpoint(path: Path, device: torch.device) -> dict:
    return torch.load(path, map_location=device, weights_only=False)
