requirements.txt
gpytorch==1.13
botorch==0.12.0
scipy==1.15.3
pytest==8.3.4
ipykernel==6.29.5

gymnasium==1.2.3
cma==4.0.0
mujoco==3.3.3
optuna==4.0.0
gymnasium[box2d]==1.2.3
gymnasium[mujoco]==1.2.3
torchrl==0.11.0
nds==0.4.3
matplotlib==3.10.8
pybind11==3.0.1
pre_commit==4.5.1
modal==1.3.0.post1
celer==0.7.4
hyperopt==0.2.7
sparse-ho @ https://github.com/QB3/sparse-ho/archive/master.zip
click==8.3.1
